+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix IDE freeze caused by duplicate background loading tasks and stuck loading state on cancellation
+- Format change ids correctly in the log
+
+### Changed
+- Reduce vertical whitespace in details panel 
 
 ## [0.3.1] - 2026-02-24